+++ src/Traits/GetPatternsHtml5AndPcre.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Estasi\Utility\Traits;
 
+use function stripslashes;
+
 /**
  * Trait GetPatternsHtml5AndPcre
  *
@@ -21,7 +23,7 @@ trait GetPatternsHtml5AndPcre
     private function getPatternsHtml5AndPCRE(string $pattern): array
     {
         if (preg_match('`^([^A-Za-z0-9\x5C\s])(.+?)\1[imsxADSUXJu]*?$`', preg_quote($pattern), $match)) {
-            return ['pcre' => $pattern, 'html' => $match[2]];
+            return ['pcre' => $pattern, 'html' => stripslashes($match[2])];
         }
 
         return ['pcre' => sprintf('`%s`', $pattern), 'html' => $pattern];