+++ CHANGELOG.md
@@ -4,8 +4,15 @@
 
 ### Added
 
-- `Traits\getPatternsHtml5AndPCRE`
+- `Interfaces\TimePeriods`
+- `Traits\GetPatternsHtml5AndPCRE`
+- `Traits\Properties__get`
+- Method `ArrayUtils::oneToMultiDimArray`
 
 ### Deprecated
 
 - `Traits\ConvertPatternHtml5ToPCRE`
+
+### Fixed
+
+- Fixed error in `ArrayUtils::iteratorToArray` (nested objects `\Traversable` did not convert to an array)